use binary tokens
